Fix multiple notifications appearing in Firefox.
Multiple notifications are being displayed in Firefox IFRAMEs.
This is a bugs where product scraper is appearing to return a valid
product for malformed HTML. This product has an 'undefined' product
code which is being treated as valid (hence the notification for "ZHE
[NOUN] Undefined").

Fixed HTML document parsing to be more strict.